toolkit/mozapps/extensions/test/mochitest/test_bug887098.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=887098
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 887098</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">
  /* exported loaded */

  /** Test for Bug 887098 **/
  SimpleTest.waitForExplicitFinish();
  /* globals $,evalRef */

  function loaded() {
    var iwin = $("ifr").contentWindow;
    var href = SpecialPowers.wrap(iwin).location.href;
    if (/file_empty/.test(href)) {
      window.evalRef = iwin.eval;
      window.installTriggerRef = iwin.InstallTrigger; // Force lazy instantiation.
      // about:mozilla is privileged, so we need to be privileged to load it.
      SpecialPowers.wrap(iwin).location.href = "about:mozilla";
    } else {
      is(href, "about:mozilla", "Successfully navigated to about:mozilla");
      try {
        evalRef('InstallTrigger.install({URL: "chrome://global/skin/global.css"});');
        ok(false, "Should have thrown when trying to install restricted URI from InstallTrigger");
      } catch (e) {
        // XXXgijs this test broke because of the switch to webidl. I'm told
        // it has to do with compartments and the fact that we eval in "about:mozilla".
        // Tracking in bug 1007671
        todo(/permission/.test(e), "We should throw a security exception. Got: " + e);
      }
      SimpleTest.finish();
    }
  }

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=887098">Mozilla Bug 887098</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<iframe onload="loaded();" id="ifr" src="file_empty.html"></iframe>
<pre id="test">
</pre>
</body>
</html>
